Create runserverpractice.py
I thought I might as well upload this file that I have been working with for the flask server.  It was helpful for me to work with all of it in one file, even though it might not be as efficient.  It was when I put it all of it into the same file that the flask server started to make more sense.